refactor(relay): derive providers from routing config
